Only try to delete dir if it exists

+++ src/libs/helpers/pathHelper.ts
@@ -79,16 +79,21 @@ export default class PathHelper {
   }
 
   /**
-   * Recursively remove everything inside a directory and then remove the directory itself.
+   * Recursively remove everything inside a
+   * directory and then remove the directory itself.
    * @param path
    */
   public static removeDir(path: string) {
-    fs.readdir(path, (_, files) => {
-      files.forEach((file) => {
-        fs.unlinkSync(Path.join(path, file));
-      });
+    if (fs.existsSync(path)) {
+      fs.readdir(path, (_, files) => {
+        if (files) {
+          files.forEach((file) => {
+            fs.unlinkSync(Path.join(path, file));
+          });
+        }
 
-      fs.rmdirSync(path);
-    });
+        fs.rmdirSync(path);
+      });
+    }
   }
 }
